backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# AI Models and ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=10.0.0
numpy>=1.24.0

# Text-to-Speech
TTS>=0.20.0
soundfile>=0.12.0
librosa>=0.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0 